+++ resolwe/flow/executors/docker/constants.py
@@ -1,0 +1,5 @@
+"""Constants for the Docker executor."""
+DATA_VOLUME = '/data'
+DATA_ALL_VOLUME = '/data_all'
+UPLOAD_VOLUME = '/upload'
+SECRETS_VOLUME = '/secrets'